Enable pkg/cpuid support on arm64.
Fixes #1255
